+++ Sources/RInAppMessaging/CampaignDispatcher.swift
@@ -108,7 +108,8 @@ internal class CampaignDispatcher: CampaignDispatcherType, TaskSchedulable {
     
     func fetchCampaignImagesAndDisplay(campaign: Campaign) {
         if let carouselData = campaign.data.customJson?.carousel,
-           !(carouselData.images?.isEmpty ?? true) {
+           !(carouselData.images?.isEmpty ?? true),
+           RInAppMessaging.isRMCEnvironment {
             fetchImagesArray(from: carouselData) { images in
                 guard let carouselData = carouselData.images else { return }
                 let carouselHandler = self.createCarouselDataList(from: carouselData, using: images)